refactor: 统一信用分字段类型并优化代码
- 将CreditScoreVO中的score、qualScore、caseScore、evalScore字段类型从Integer改为Byte以匹配实体类型
- 优化EvaluationServiceImpl中匿名评价的布尔值转换逻辑
- 重构CreditController.convertToVO方法，直接赋值替代BeanUtils.copyProperties

+++ zhilian-backend/src/main/java/com/zhilian/zhilianbackend/dto/response/CreditScoreVO.java
@@ -21,16 +21,16 @@ public class CreditScoreVO {
     private Long serviceId;
 
     @Schema(description = "综合信用分")
-    private Integer score;
+    private Byte score;
 
     @Schema(description = "资质分")
-    private Integer qualScore;
+    private Byte qualScore;
 
     @Schema(description = "案例分")
-    private Integer caseScore;
+    private Byte caseScore;
 
     @Schema(description = "评价分")
-    private Integer evalScore;
+    private Byte evalScore;
 
     @Schema(description = "计算时间")
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
